Wrote Code for Reading two Ultrasonic Sensors
I added interrupts to watch for ultrasonic signal pulses. When a pulse
goes HIGH, a note of the time is taken (in micro-seconds). When the same
pulse goes LOW, the time difference is added to a rolling array of time
values. The function read_dist() averages the current 10 distance values
from both front and back and converts them into a singer averaged
distance which is returned as a float.