Added PWM to trigger Ultrasonic sensors.
There are two ultrasonic outputs. They are on pins 4 and 5 (OC2 and
OC3). These have periods of 50ms and on times of 10us. The periods are
offset by 25ms, with OC2 leading OC3. This allows the ultrasonics to run
at different times and not interfere with each other. They will run in
the background and the echo will trigger an interrupt (already
committed). The interrupt will update an array of distance measurements
every 50ms. This continually updated array of distances will be sampled
for navigation purposes.