Cleaned the Board Files
Removed all outdated functions unless the algorithms could still be
useful, in which case I just block commented it. Also, realized that
there was a mistake in the find_normal() function. In stead of checking
a value, I assigned it to the value I was comparing it to, the old "="
vs "==" problem. It's fixed now. Everything compiles, ultrasonic,
stepping, normal, and competition...