Added stage_timer() and stage_change() functions (UN-COMPILED)
I added means to track the time within the stage. This will help us
change target if the 30 sec timer expires. It's not yet been compiled to
verify if works, but It's a start.

TO DO: Write code to track how many shots have been taken, finish
tracking which targets have been shot into, verify find_normal() works
and if it doesn't, check the find_corner() function. Verify ADC on the
selected pins....