Wrote aquire_target() + some others
This now intelligently decides where to check for IR. It might not work
(hardware is untested), but it compiles. It tracks when the right
(bit0), striaght (bit1), and left (bit2) direction IR beacons are found.
If the beacon was found in a prior time stage, the robot doesn't turn
that way to check. Otherwise, it looks straight, right, then left. If it
can't find it anywhere, it will just start turning 90 degrees at a time
and check regardless of if it's "been found" there before.